+++ typescript/infra/scripts/helloworld/kathy.ts
@@ -17,15 +17,12 @@ import {
 import {
   Address,
   ProtocolType,
-  debug,
   ensure0x,
-  error,
-  log,
   objMap,
   retryAsync,
+  rootLogger,
   strip0x,
   timeout,
-  warn,
 } from '@hyperlane-xyz/utils';
 
 import { Contexts } from '../../config/contexts';
@@ -44,6 +41,8 @@ import { getArgs, withContext } from '../agent-utils';
 import { getEnvironmentConfig } from '../core-utils';
 
 import { getHelloWorldMultiProtocolApp } from './utils';
+
+const logger = rootLogger.child({ module: 'kathy' });
 
 const metricsRegister = new Registry();
 // TODO rename counter names
@@ -162,7 +161,7 @@ async function main(): Promise<boolean> {
   let errorOccurred = false;
 
   startMetricsServer(metricsRegister);
-  debug('Starting up', { environment });
+  logger.debug('Starting up', { environment });
 
   const coreConfig = getEnvironmentConfig(environment);
   // const coreConfig = getCoreConfigStub(environment);
@@ -202,7 +201,7 @@ async function main(): Promise<boolean> {
     .filter((v) => v !== null)
     .map((v) => v!);
 
-  debug('Pairings calculated', { chains, pairings });
+  logger.debug('Pairings calculated', { chains, pairings });
 
   let allowedToSend: number;
   let currentPairingIndex: number;
@@ -214,7 +213,7 @@ async function main(): Promise<boolean> {
     // Start with pairing 0
     currentPairingIndex = 0;
 
-    debug('Cycling once through all pairs');
+    logger.debug('Cycling once through all pairs');
   } else {
     // If we are not cycling just once and are running this as a service, do so at an interval.
     // Track how many we are still allowed to send in case some messages send slower than expected.
@@ -223,14 +222,14 @@ async function main(): Promise<boolean> {
     // in case we are restarting kathy, keep it from always running the exact same messages first
     currentPairingIndex = Date.now() % pairings.length;
 
-    debug('Running as a service', {
+    logger.debug('Running as a service', {
       sendFrequency,
     });
 
     setInterval(() => {
       // bucket cap since if we are getting really behind it probably does not make sense to let it run away.
       allowedToSend = Math.min(allowedToSend + 1, MAX_MESSAGES_ALLOWED_TO_SEND);
-      debug('Tick; allowed to send another message', {
+      logger.debug('Tick; allowed to send another message', {
         allowedToSend,
         sendFrequency,
       });
@@ -265,7 +264,7 @@ async function main(): Promise<boolean> {
         await app.stats(),
       )) {
         for (const [destination, counts] of Object.entries(destinationStats)) {
-          debug('Message stats', {
+          logger.debug('Message stats', {
             origin,
             destination,
             currentCycle,
@@ -278,7 +277,7 @@ async function main(): Promise<boolean> {
       cycleMessageCount = 0;
 
       if (cycleOnce) {
-        log('Finished cycling through all pairs once');
+        logger.info('Finished cycling through all pairs once');
         // Return true to signify messages should stop being sent.
         return true;
       }
@@ -310,7 +309,7 @@ async function main(): Promise<boolean> {
       (origin === 'ethereum' || destination === 'ethereum') &&
       currentCycle % (cyclesBetweenEthereumMessages + 1) !== 0
     ) {
-      debug('Skipping message to/from Ethereum', {
+      logger.debug('Skipping message to/from Ethereum', {
         currentCycle,
         origin,
         destination,
@@ -334,7 +333,7 @@ async function main(): Promise<boolean> {
     // In the cycle-once case, the loop is expected to exit before ever hitting
     // this condition.
     if (allowedToSend <= 0) {
-      debug('Waiting before sending next message', {
+      logger.debug('Waiting before sending next message', {
         ...logCtx,
         sendFrequency,
       });
@@ -342,7 +341,7 @@ async function main(): Promise<boolean> {
     }
     allowedToSend--;
 
-    debug('Initiating sending of new message', logCtx);
+    logger.debug('Initiating sending of new message', logCtx);
 
     try {
       await sendMessage(
@@ -356,10 +355,10 @@ async function main(): Promise<boolean> {
         messageSendTimeout,
         messageReceiptTimeout,
       );
-      log('Message sent successfully', { origin, destination });
+      logger.info('Message sent successfully', { origin, destination });
       messagesSendCount.labels({ ...labels, status: 'success' }).inc();
     } catch (e) {
-      error(`Error sending message, continuing...`, {
+      logger.error(`Error sending message, continuing...`, {
         error: format(e),
         ...logCtx,
       });
@@ -371,7 +370,7 @@ async function main(): Promise<boolean> {
       origin,
       coreConfig.owners[origin].owner,
     ).catch((e) => {
-      warn('Failed to update wallet balance for chain', {
+      logger.warn('Failed to update wallet balance for chain', {
         chain: origin,
         err: format(e),
       });
@@ -419,7 +418,7 @@ async function sendMessage(
 
   const metricLabels = { origin, remote: destination };
 
-  log('Sending message', {
+  logger.info('Sending message', {
     origin,
     destination,
     interchainGasPayment: value,
@@ -493,7 +492,7 @@ async function sendMessage(
   );
   messageSendSeconds.labels(metricLabels).inc((Date.now() - startTime) / 1000);
 
-  log('Message sent, waiting for it to be processed', {
+  logger.info('Message sent, waiting for it to be processed', {
     origin,
     destination,
     receipt,
@@ -509,7 +508,7 @@ async function sendMessage(
   messageReceiptSeconds
     .labels(metricLabels)
     .inc((Date.now() - startTime) / 1000);
-  log('Message received', {
+  logger.info('Message received', {
     origin,
     destination,
   });
@@ -535,7 +534,11 @@ async function updateWalletBalanceMetricFor(
       token_symbol: 'Native',
     })
     .set(balance);
-  debug('Wallet balance updated for chain', { chain, signerAddress, balance });
+  logger.debug('Wallet balance updated for chain', {
+    chain,
+    signerAddress,
+    balance,
+  });
 }
 
 // Get a core config intended for testing Kathy without secret access
@@ -549,7 +552,7 @@ export function getCoreConfigStub(environment: DeployEnvironment) {
   const privateKeyEvm = process.env.KATHY_PRIVATE_KEY_EVM;
   if (!privateKeyEvm) throw new Error('KATHY_PRIVATE_KEY_EVM env var not set');
   const evmSigner = new Wallet(privateKeyEvm);
-  console.log('evmSigner address', evmSigner.address);
+  logger.info('evmSigner address', evmSigner.address);
   multiProvider.setSharedSigner(evmSigner);
 
   // const privateKeySealevel = process.env.KATHY_PRIVATE_KEY_SEALEVEL;
@@ -559,7 +562,7 @@ export function getCoreConfigStub(environment: DeployEnvironment) {
   // const sealevelSigner = Keypair.fromSeed(
   //   Buffer.from(privateKeySealevel, 'hex'),
   // );
-  // console.log('sealevelSigner address', sealevelSigner.publicKey.toBase58());
+  // console.logger.info('sealevelSigner address', sealevelSigner.publicKey.toBase58());
 
   const testnetKeys = objMap(testnetConfigs, (_, __) => ({
     address: evmSigner.address,
@@ -580,15 +583,15 @@ export function getCoreConfigStub(environment: DeployEnvironment) {
 
 main()
   .then((errorOccurred: boolean) => {
-    log('Main exited');
+    logger.info('Main exited');
     if (errorOccurred) {
-      error('An error occurred at some point');
+      logger.error('An error occurred at some point');
       process.exit(1);
     } else {
       process.exit(0);
     }
   })
   .catch((e) => {
-    error('Error in main', { error: format(e) });
+    logger.error('Error in main', { error: format(e) });
     process.exit(1);
   });